Fix error when cancelling loading of json wallet. Ref: DGDG-222
The error happens because we call `props.onClose` whenever we reset the modal. However, `props.onClose` updates the state of the wallet in governance-ui-components, so the app thinks the wallet has successfully loaded the keystore when it should have actually reset. This diff splits `onClose` to `onSuccess` and `onReset` and calls these methods in the proper stages of the load wallet lifecycle.