[FIX] Make sure ledger gets updated gas price when signing transactions
Ref: DGDG-470

Bumps `react-ledger-container` to `v0.1.8` and fixes issue with ledger wallets not using the updated gas price in transactions.

Previously, the transaction modal will immediately initiate the transaction upon rendering. Changes in props will not propagate to `<LedgerContainer/>` once this happens, since the component already created a promise that waits for the transaction to be signed by the user. This leads to the bug where, if the user edits the gas price before signing, the new gas price isn't used in the transaction because `<LedgerContainer/>` has already been initialized with the default gas price. `v0.1.8` of `react-ledger-container` should fix this issue by requiring the `renderInitSigning` prop that allows us to manually initiate the transaction from the UI side.

The new UI behavior is detailed in the test plan below.

**Test Plan**
- Load a ledger wallet and try to lock some DGD.
- Change the gas price via the **Advanced** tab.
- The transaction modal should have an additional button that initiates the transaction signing. Once you click this button, the **Advanced** tab should disappear and your ledger device will ask you to sign the transaction.
- Sign the transaction and take note of the `txHash` from the resulting snackbar message (or from the Transaction History page).
- You can check if the gas price of the transaction was correct by running the following command on `truffle console`: `web3.eth.getTransaction([txHash])` where `[txHash]` is the hash enclosed in single-quotes.