[FIX] Remove Cancel Signing button in UI when confirmation is asked from non-json wallets
We don't have a way to cancel the transaction from the UI once it's being asked to confirm on a ledger, trezor, or metamask device. In these cases, we should just remove the button from the UI and let the user cancel from their device.

**Test Plan**

Try to load a wallet and lock DGD using different kinds of wallets.
- JSON wallets should still work as is (Cancel Signing button should still be present).
- Ledger wallets will not have the Cancel Signing button once you click Sign Transaction. User can cancel from their device.
- Metamask should not have the Cancel Signing button or the Advanced tab. You can reject a transaction and change the gas fee from the metamask window, so having these in the UI would be redundant.