Feature/sprint 7 (part a) (#20)
* add scripts to start local UI for kovan/production

* updated zendesk key and color scheme

* updated reference to react-trezor-library; updated trezor singning components

* [FIX] Make sure ledger gets updated gas price when signing transactions

Ref: DGDG-470

Bumps `react-ledger-container` to `v0.1.8` and fixes issue with ledger wallets not using the updated gas price in transactions.

Previously, the transaction modal will immediately initiate the transaction upon rendering. Changes in props will not propagate to `<LedgerContainer/>` once this happens, since the component already created a promise that waits for the transaction to be signed by the user. This leads to the bug where, if the user edits the gas price before signing, the new gas price isn't used in the transaction because `<LedgerContainer/>` has already been initialized with the default gas price. `v0.1.8` of `react-ledger-container` should fix this issue by requiring the `renderInitSigning` prop that allows us to manually initiate the transaction from the UI side.

The new UI behavior is detailed in the test plan below.

**Test Plan**
- Load a ledger wallet and try to lock some DGD.
- Change the gas price via the **Advanced** tab.
- The transaction modal should have an additional button that initiates the transaction signing. Once you click this button, the **Advanced** tab should disappear and your ledger device will ask you to sign the transaction.
- Sign the transaction and take note of the `txHash` from the resulting snackbar message (or from the Transaction History page).
- You can check if the gas price of the transaction was correct by running the following command on `truffle console`: `web3.eth.getTransaction([txHash])` where `[txHash]` is the hash enclosed in single-quotes.

* added error handling for transaction cancellation using trezor

* return reject when error is encountered when trying to add an address using metamask

* [FIX] Remove Cancel Signing button in UI when confirmation is asked from non-json wallets

We don't have a way to cancel the transaction from the UI once it's being asked to confirm on a ledger, trezor, or metamask device. In these cases, we should just remove the button from the UI and let the user cancel from their device.

**Test Plan**

Try to load a wallet and lock DGD using different kinds of wallets.
- JSON wallets should still work as is (Cancel Signing button should still be present).
- Ledger wallets will not have the Cancel Signing button once you click Sign Transaction. User can cancel from their device.
- Metamask should not have the Cancel Signing button or the Advanced tab. You can reject a transaction and change the gas fee from the metamask window, so having these in the UI would be redundant.

* Added initial transaction message so user's can perform transaction on their own

* bumped react-trezor-container version

* typo fix

* [IMPROVEMENT] Remove console errors for json and ledger wallets

**Test Plan**

Test that loading a wallet and signing a transaction still works.
- Load a json wallet and lock DGD. It should work without console errors.
- Load a ledger wallet and lock DGD. It should work without console errors.

* dynamically pulls the contract address of DGD contract to display balances when loading wallets for Trezor and Ledger

* added special handling on dgd for displaying balance

* [FIX] Fix issue with loading metamask wallet

There was a breaking change in Metamask late last year (see: https://medium.com/metamask/https-medium-com-metamask-breaking-change-injecting-web3-7722797916a8). This diff should cover the changes needed.

**Test Plan**
- Install the latest version of Metamask.
- Try to load your Metamask wallet. You should receive no errors.

* Updated handling of keystores so metamask wallet is handled differently and reverted the behavior for other wallets as before

* update on showing balance for Trezor Address